requirements.txt
alembic==1.8.1
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
Flask==2.2.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
gunicorn==20.1.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
mysqlclient==2.1.1
packaging==21.3
pycparser==2.21
pyparsing==3.0.9
python-dotenv==0.21.0
six==1.16.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2
zipp==3.8.1
